--- src/components/Advert.js
@@ -1,4 +1,0 @@
-import React from 'react';
-
-
-export {Advert};
