--- src/components/Banner.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export {Banner};
