--- src/components/CategoriesList.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export {CategoriesList}
