--- src/components/Header.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export {Header};
